Change the way you capture frames from video
- Instead of FFmpeg, MediaMetadataRetriever was used to get frames
- Improved performance of the method of creating collages from video frames

+++ app/src/main/java/com/project/videoeditor/VideoInfo.java
@@ -156,8 +156,8 @@ public class VideoInfo implements Parcelable {
         return aspectRatio;
     }
 
-    public Long getDuration() {
-        return duration;
+    public int getDuration() {
+        return Math.toIntExact(duration);
     }
 
     public Long getBitrate() {
